+++ src/index.ts
@@ -3,4 +3,3 @@ export * from "./viewport/index";
 export * from "./config";
 export * from "./module";
 export * from "./version";
-export * from "./platform/window";