+++ src/config.ts
@@ -1,14 +1,17 @@
 import { InjectionToken } from "@angular/core";
-import { UxViewportOptions } from "./viewport/viewport.model";
+import { UX_VIEWPORT_DEFAULT_CONFIG } from './viewport/viewport.const';
+import { UxViewportOptions, UxViewportOptionsInternal } from "./viewport/viewport.model";
 
 export interface UxOptions {
-	viewport: Partial<UxViewportOptions>;
+	viewport: UxViewportOptions;
 }
 
 export const UX_DEFAULT_CONFIG: UxOptions = {
-	viewport: {
-		resizePollingSpeed: 33,
-	}
+	viewport: UX_VIEWPORT_DEFAULT_CONFIG
 };
+
+export interface UxOptionsInternal {
+	viewport: UxViewportOptionsInternal;
+}
 
 export const UX_CONFIG = new InjectionToken<UxOptions>("@ssv/ngx.ux-config");