+++ src/module.ts
@@ -1,7 +1,8 @@
+import * as _ from "lodash";
 import { NgModule, ModuleWithProviders, InjectionToken } from "@angular/core";
 
-import { SsvViewportMatcherDirective } from "./viewport/index";
-import { UxOptions, UX_DEFAULT_CONFIG, UX_CONFIG } from "./config";
+import { generateViewportInternalConfig, SsvViewportMatcherDirective } from "./viewport/index";
+import { UxOptions, UX_DEFAULT_CONFIG, UX_CONFIG, UxOptionsInternal } from "./config";
 import { WINDOW } from "./platform/window";
 
 /** @internal */
@@ -13,7 +14,6 @@ export const _MODULE_CONFIG = new InjectionToken<UxOptions | (() => UxOptions)>(
 	declarations: [SsvViewportMatcherDirective],
 	providers: [
 		{ provide: UX_CONFIG, useValue: UX_DEFAULT_CONFIG },
-
 		{ provide: WINDOW, useFactory: _window },
 	],
 	exports: [SsvViewportMatcherDirective],
@@ -28,15 +28,21 @@ export class SsvUxModule {
 					useFactory: _moduleConfigFactory,
 					deps: [_MODULE_CONFIG],
 				},
-				{ provide: _MODULE_CONFIG, useValue: config || UX_DEFAULT_CONFIG },
+				{ provide: _MODULE_CONFIG, useValue: config },
 			],
 		};
 	}
 }
 
 /** @internal */
-export function _moduleConfigFactory(config: UxOptions | (() => UxOptions)) {
-	return typeof config === "function" ? config() : config;
+export function _moduleConfigFactory(config: UxOptions | (() => UxOptions)): UxOptionsInternal {
+	const value = typeof config === "function" ? config() : config;
+	const mergedValue = _.merge(UX_DEFAULT_CONFIG, value);
+	const viewport = generateViewportInternalConfig(mergedValue.viewport);
+
+	return {
+		viewport
+	};
 }
 
 /** @internal */