+++ src/viewport/viewport.service.ts
@@ -11,63 +11,10 @@ import {
 	auditTime,
 } from "rxjs/operators";
 
-import { Dictionary } from "../internal/internal.model";
-import { UxOptions, UX_CONFIG } from "../config";
+import { UxOptionsInternal, UX_CONFIG } from "../config";
 import { ViewportSizeTypeInfo, ViewportSize, ViewportSizeType } from "./viewport.model";
 import { WindowRef } from "../platform/window";
 import { ViewportServerSizeService } from "./viewport-server-size.service";
-import { UX_VIEWPORT_DEFAULT_CONFIG } from "./viewport.const";
-
-// todo: make this configurable
-/** Viewport sizes config, by upper bound. e.g. given width '1000' and `medium` is set to '992' => `large`. */
-const viewportSizesConfig = {
-	xsmall: 450,
-	small: 767,
-	medium: 992,
-	large: 1200,
-	xlarge: 1500,
-	xxlarge: 1920,
-	xxlarge1: 2100
-};
-
-// todo: autogenereate
-const viewportSizeRefs: Dictionary<Readonly<ViewportSizeTypeInfo>> = {
-	[ViewportSizeType.xsmall]: Object.freeze({
-		name: "xsmall",
-		type: ViewportSizeType.xsmall,
-		widthThreshold: viewportSizesConfig.xsmall,
-	} as ViewportSizeTypeInfo),
-	[ViewportSizeType.small]: Object.freeze({
-		name: "small",
-		type: ViewportSizeType.small,
-		widthThreshold: viewportSizesConfig.small,
-	} as ViewportSizeTypeInfo),
-	[ViewportSizeType.medium]: Object.freeze({
-		name: "medium",
-		type: ViewportSizeType.medium,
-		widthThreshold: viewportSizesConfig.medium,
-	} as ViewportSizeTypeInfo),
-	[ViewportSizeType.large]: Object.freeze({
-		name: "large",
-		type: ViewportSizeType.large,
-		widthThreshold: viewportSizesConfig.large,
-	} as ViewportSizeTypeInfo),
-	[ViewportSizeType.xlarge]: Object.freeze({
-		name: "xlarge",
-		type: ViewportSizeType.xlarge,
-		widthThreshold: viewportSizesConfig.xlarge,
-	} as ViewportSizeTypeInfo),
-	[ViewportSizeType.xxlarge]: Object.freeze({
-		name: "xxlarge",
-		type: ViewportSizeType.xxlarge,
-		widthThreshold: viewportSizesConfig.xxlarge,
-	} as ViewportSizeTypeInfo),
-	[ViewportSizeType.xxlarge1]: Object.freeze({
-		name: "xxlarge1",
-		type: ViewportSizeType.xxlarge1,
-		widthThreshold: viewportSizesConfig.xxlarge1,
-	} as ViewportSizeTypeInfo)
-};
 
 @Injectable({
 	providedIn: "root",
@@ -83,7 +30,7 @@ export class ViewportService {
 	private lastWidthSizeInfo: ViewportSizeTypeInfo | undefined;
 
 	constructor(
-		@Inject(UX_CONFIG) config: UxOptions,
+		@Inject(UX_CONFIG) private config: UxOptionsInternal,
 		@Inject(DOCUMENT) private document: any,
 		private windowRef: WindowRef,
 		private viewportServerSize: ViewportServerSizeService,
@@ -91,7 +38,7 @@ export class ViewportService {
 		if (windowRef.hasNative) {
 			this.resize$ = fromEvent<Event>(window, "resize").pipe(
 				map(() => this.getViewportSize()),
-				auditTime(config.viewport.resizePollingSpeed || UX_VIEWPORT_DEFAULT_CONFIG.resizePollingSpeed),
+				auditTime(this.config.viewport.resizePollingSpeed),
 				share(),
 			);
 		} else {
@@ -145,27 +92,34 @@ export class ViewportService {
 	}
 
 	private calculateViewportSize(width: number): ViewportSizeTypeInfo {
-		if (width === this.lastWidthCheck && this.lastWidthSizeInfo) {
-			return this.lastWidthSizeInfo;
+		if (width !== this.lastWidthCheck) {
+			this.lastWidthCheck = width;
+			this.lastWidthSizeInfo = this.getWidthSizeInfo(width);
 		}
-		this.lastWidthCheck = width;
 
-		// todo: make this more dynamic + decouple from lodash
-		if (_.inRange(width, viewportSizesConfig.xsmall)) {
-			this.lastWidthSizeInfo = viewportSizeRefs[ViewportSizeType.xsmall];
-		} else if (_.inRange(width, viewportSizesConfig.xsmall, viewportSizesConfig.small)) {
-			this.lastWidthSizeInfo = viewportSizeRefs[ViewportSizeType.small];
-		} else if (_.inRange(width, viewportSizesConfig.small, viewportSizesConfig.medium)) {
-			this.lastWidthSizeInfo = viewportSizeRefs[ViewportSizeType.medium];
-		} else if (_.inRange(width, viewportSizesConfig.medium, viewportSizesConfig.large)) {
-			this.lastWidthSizeInfo = viewportSizeRefs[ViewportSizeType.large];
-		} else if (_.inRange(width, viewportSizesConfig.large, viewportSizesConfig.xlarge)) {
-			this.lastWidthSizeInfo = viewportSizeRefs[ViewportSizeType.xlarge];
-		} else if (_.inRange(width, viewportSizesConfig.xlarge, viewportSizesConfig.xxlarge)) {
-			this.lastWidthSizeInfo = viewportSizeRefs[ViewportSizeType.xxlarge];
-		} else {
-			this.lastWidthSizeInfo = viewportSizeRefs[ViewportSizeType.xxlarge1];
-		}
 		return this.lastWidthSizeInfo;
+	}
+
+	/**
+	 * Returns the current viewport information
+	 * @param width the viewport width
+	 */
+	private getWidthSizeInfo(width: number): ViewportSizeTypeInfo {
+		// todo: make this more dynamic + decouple from lodash
+		if (_.inRange(width, this.config.viewport.breakpoints.small)) {
+			return this.config.viewport.viewportDictionary[ViewportSizeType.xsmall];
+		} else if (_.inRange(width, this.config.viewport.breakpoints.xsmall, this.config.viewport.breakpoints.small)) {
+			return this.config.viewport.viewportDictionary[ViewportSizeType.small];
+		} else if (_.inRange(width, this.config.viewport.breakpoints.small, this.config.viewport.breakpoints.medium)) {
+			return this.config.viewport.viewportDictionary[ViewportSizeType.medium];
+		} else if (_.inRange(width, this.config.viewport.breakpoints.medium, this.config.viewport.breakpoints.large)) {
+			return this.config.viewport.viewportDictionary[ViewportSizeType.large];
+		} else if (_.inRange(width, this.config.viewport.breakpoints.large, this.config.viewport.breakpoints.xlarge)) {
+			return this.config.viewport.viewportDictionary[ViewportSizeType.xlarge];
+		} else if (_.inRange(width, this.config.viewport.breakpoints.xlarge, this.config.viewport.breakpoints.xxlarge)) {
+			return this.config.viewport.viewportDictionary[ViewportSizeType.xxlarge];
+		}
+
+		return this.config.viewport.viewportDictionary[ViewportSizeType.xxlarge1];
 	}
 }