+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Features
 
-- **viewport:** changes to `ViewportService` to be overriden by changing `getViewportSize` and `calculateViewportSize` to protected
+- **viewport:** changes to `ViewportService` to be overriden by changing `getViewportSize` & `calculateViewportSize` to protected and expose `pollingSpeed`
 - **viewport:** export `WindowRef`
 
 ## [1.1.1](https://github.com/sketch7/ngx.ux/compare/1.1.0...1.1.1) (2020-05-29)