+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.1.2](https://github.com/sketch7/ngx.ux/compare/1.1.1...1.1.2) (2020-10-24)
+
+### Features
+
+- **viewport:** add configurable breakpoints
+
 ## [1.1.1](https://github.com/sketch7/ngx.ux/compare/1.1.0...1.1.1) (2020-05-29)
 
 ### Bug Fixes